analytics_service/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
PyJWT>=2.8,<3.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# HTTP Client (for aggregating data)
requests>=2.31,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Security
python-dotenv>=1.0,<2.0

# WSGI Server
gunicorn>=21.2,<22.0
